--- v8py/utils.h
@@ -1,8 +1,0 @@
-#ifndef UTILS_H
-#define UTILS_H
-
-#include <Python.h>
-
-PyObject *construct_new_object(PyObject *self, PyObject *args);
-
-#endif
